Common FGElement base class
